requirements.txt
prefect==2.19.9
PyGithub==2.4.0
python-gitlab==4.12.0
#numpy==2.1.2
numpy
#pandas==2.2.3

requests==2.32.3
pyyaml==6.0.2
blake3==1.0.4
datasketch==1.6.5
jsonschema==4.23.0
GitPython==3.1.43
python-dotenv==1.0.1
halo==0.0.31
rich==13.9.2
orjson==3.10.7
fasteners==0.19
pydantic==2.11.1
typing-extensions>=4.9.0
# Security tools are invoked via Docker CLIs (Semgrep, Bandit, Gitleaks, ScanCode)

pytest==8.3.3
black
radon
